fix: refactor our boilerplate
